+++ matrix_actions.h
@@ -1,7 +1,9 @@
 #ifndef MATRIX_ACTIONS_H
 #define MATRIX_ACTIONS_H
 
-void transpose(double* matrix, const int& rows, const int & cols);
-void oddEvenRows(double* matrix, const int& rows, const int& cols);
+#include <cstddef>
+
+void transpose(double* matrix, const size_t& rows, const size_t& cols);
+void oddEvenRows(double* matrix, const size_t& rows, const size_t& cols);
 
 #endif //MATRIX_ACTIONS_H
