+++ matrixio.h
@@ -1,7 +1,9 @@
 #ifndef MATRIXIO_H
 #define MATRIXIO_H
 
-void matrixin(double* matrix, const int& rows, const int& cols);
-void matrixout(double* matrix, const int& rows, const int& cols);
+#include <cstddef>
+
+void matrixin(double* matrix, const size_t& rows, const size_t& cols);
+void matrixout(double* matrix, const size_t& rows, const size_t& cols);
 
 #endif //MATRIXIO_H
